backend/requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
appdirs==1.4.4
black==21.4b2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.122.0
filetype==1.2.0
fonttools==4.60.1
future==1.0.0
fvcore==0.1.5.post20221221
grpcio==1.76.0
h11==0.16.0
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.9
kiwisolver==1.4.7
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
mkl-service==2.4.0
mypy_extensions==1.1.0
numpy==1.24.4
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
packaging==25.0
pathspec==0.12.1
pi_heif==1.1.1
Pillow
pillow-avif-plugin==1.5.2
portalocker==3.2.0
protobuf==6.33.1
pycocotools==2.0.10
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
six==1.17.0
starlette==0.49.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
yacs==0.1.8
zipp==3.23.0
